requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-cache2==0.2.1
flower==2.0.1
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
pendulum==2.1.2
Pillow==10.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.1
SQLAlchemy==2.0.21
starlette==0.27.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
WTForms==3.1.0
zipp==3.17.0

pytest~=7.4.2